requirements.txt
langchain-google-genai
langchain
python-dotenv
langchain-community
langchain-huggingface
accelerate
unstructured
pdfminer
pdfminer.six
pi_heif
unstructured_inference
pdf2image
fastapi[standard]
python-multipart
